requirements_dev.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
Django==1.9.6
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
gnureadline==6.3.3
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
Markdown==2.6.6
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
pyreadline==2.1
PyYAML==3.11
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
